+++ src/main/kotlin/com/chatandpay/api/dto/OtherBankAccountRequestDTO.kt
@@ -1,4 +1,4 @@
-package com.chatandpay.api.controller.dto
+package com.chatandpay.api.dto
 
 data class OtherBankAccountRequestDTO(
     val bankCode: String,
