+++ src/main/kotlin/com/chatandpay/api/dto/SignUpPayUserDTO.kt
@@ -1,4 +1,4 @@
-package com.chatandpay.api.controller.dto
+package com.chatandpay.api.dto
 
 data class SignUpPayUserDTO(
     val ci: String,
